refactor: modal 2가지 유형 정의
